init: skeleton of project with basic plugin setup
Problem: we need an easy interface for developers to write plugins!
A plugin can be an extractor or (TBA) other. Generally speaking, it
will run a tool of interest and map the output (e.g., application
performance or similar) into a namespace defined for a compatibility
specification. The calling library (compspec) discovers the plugin
and knows how to handle validation and execution of these steps
along with packaging of the metadata. It is going to be neat!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>